+++ index.js
@@ -9,6 +9,7 @@ function requireScript(scriptId) {
         case 'common': return require('./src/jsu_common.js');
         case 'csv-parser': return require('./src/jsu_csv_parser.js');
         case 'event': return require('./src/jsu_event.js');
+        case 'latex': return require('./src/jsu_latex.js');
     }
     throw new Error('Failed to load script id: ' + scriptId);
 }
@@ -21,4 +22,5 @@ module.exports = {
     get Common() { return requireScript('common'); },
     get CsvParser() { return requireScript('csv-parser'); },
     get Event() { return requireScript('event'); },
+    get Latex() { return requireScript('latex'); },
 };