+++ tests/_index.js
@@ -14,7 +14,7 @@ const assert = require('assert');
     (function() {
         describe('requiring the package', () => {
             it('should only expose the expected properties', () => {
-                const expectedProps = ['Common', 'CsvParser', 'Event'];
+                const expectedProps = ['Common', 'CsvParser', 'Event', 'Latex'];
                 const obj = Object.assign({}, Jsu);
                 for(const prop of expectedProps) delete obj[prop];
                 assert.deepStrictEqual(obj, {});
@@ -23,22 +23,30 @@ const assert = require('assert');
 
         describe('requiring scripts more than once from the package', () => {
             it('should always expose the same object', () => {
+                const k = 5;
+
                 const common = require('../src/jsu_common.js');
-                for(let i = 0; i < 5; i++) {
+                for(let i = 0; i < k; i++) {
                     assert.strictEqual(Jsu.Common, common);
                     assert.deepStrictEqual(Jsu.Common, common);
                 }
 
                 const csv_parser = require('../src/jsu_csv_parser.js');
-                for(let i = 0; i < 5; i++) {
+                for(let i = 0; i < k; i++) {
                     assert.strictEqual(Jsu.CsvParser, csv_parser);
                     assert.deepStrictEqual(Jsu.CsvParser, csv_parser);
                 }
 
                 const event = require('../src/jsu_event.js');
-                for(let i = 0; i < 5; i++) {
+                for(let i = 0; i < k; i++) {
                     assert.strictEqual(Jsu.Event, event);
                     assert.deepStrictEqual(Jsu.Event, event);
+                }
+
+                const latex = require('../src/jsu_latex.js');
+                for(let i = 0; i < k; i++) {
+                    assert.strictEqual(Jsu.Latex, latex);
+                    assert.deepStrictEqual(Jsu.Latex, latex);
                 }
             });
         });